-upd: ??????????? (r1795) ???????? ?? ?????? 3.0.2 ?? ???????????? ???????????: scintilla 09.12.2011 r3965 scite 09.12.2011 r3713

+++ src/scintilla/cocoa/InfoBar.mm
@@ -46,7 +46,7 @@
 //--------------------------------------------------------------------------------------------------
 
 - (void) selectWithFrame: (NSRect) aRect inView: (NSView*) controlView editor: (NSText*) textObj 
-                delegate:(id) anObject start: (NSInteger) selStart length: (NSInteger) selLength
+                delegate:(id) anObject start: (int) selStart length: (int) selLength
 {
 	aRect = [self drawingRectForBounds: aRect];
 	mIsEditingOrSelecting = YES;
